create class ProductTest.java with test ProductExpectedCorrectValuesByGettersWhenIntitialized. Fits.

+++ src/main/java/Product.java
@@ -1,7 +1,7 @@
 import java.math.BigDecimal;
 
 public record Product(
-        int ean,
+        long ean,
         String productName,
         String manufacturer,
         String scale,
